Add files via upload
An IT consultant from Lagadatorumeå.se go out to you and solve the problem on the spot. If the problem is too advanced to be solved on the spot, be with your computer to our service center and fixed there. After the service, the computer is back and switched on, so everything works as it should.